+++ packages/terra-data-grid/package.json
@@ -40,6 +40,7 @@
     "terra-icon": "^2.6.0",
     "terra-aggregator": "^2.4.0",
     "terra-content-container": "^2.0.1",
+    "terra-doc-template": "^1.12.0",
     "terra-overlay": "^2.16.0",
     "terra-slide-panel-manager": "^2.3.0",
     "terra-modal-manager": "^2.4.0",
