--- packages/terra-data-grid/src/terra-dev-site/doc/example/ContentCellLayout.scss
@@ -1,13 +1,0 @@
-:local {
-  .content-cell {
-    height: 100%;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    width: 100%;
-    word-break: normal;
-    word-wrap: normal;
-    padding: 8px;
-    vertical-align: center;
-  }
-}
